Update README.md
https://en.wikipedia.org/wiki/The_Open_Source_Definition

For something to be open source, it has a number of requirements, and of the core ones that are relevant, is number 6.

"No discrimination against fields of endeavor: The license must not restrict anyone from making use of the program in a specific field of endeavor. For example, it may not restrict the program from being used in a business, or from being used for genetic research."

Further, the RPL (https://opensource.org/licenses/RPL-1.5) states:

"7.1 If You create or use a modified version of this License, which You may do
only in order to apply it to software that is not already Licensed Software
under this License, You must rename Your license so that it is not confusingly
similar to this License, and must make it clear that Your license contains
terms that differ from this License. In so naming Your license, You may not
use any trademark of Licensor or of any Contributor. Should Your modifications
to this License be limited to alteration of a) Section 13.8 solely to modify
the legal Jurisdiction or Venue for disputes, b) EXHIBIT A solely to define
License Notice text, or c) to EXHIBIT B solely to define a User-Visible
Attribution Notice, You may continue to refer to Your License as the
Reciprocal Public License or simply the RPL."

I'd argue that the requirements imposed in your PDF makes it a *new* license.