Add TokenResponseGenerator.CreateResponseAsync
CreateResponseAsync is a new virtual method in the
TokenResponseGenerator. CreateResponseAsync does the same thing that
ProcessTokenRequestAsync used to do (generate access and refresh tokens
and add them to a response object), but it is called by all token flows
except the refresh token flow, which generates its tokens differently.

TokenResponseGenerator.ProcessTokenRequestAsync is now an alias for
CreateResponseAsync and is deprecated in favor of it, with the intention
to completely remove ProcessTokenRequestAsync in 7.0.