Make ConfigurationDbContext overridable
To inherit from ConfigurationContext it is easier to use the base class of DbContextOptions without generic like already done in the PersistedGrantDbContext. This makes it easer to add additional DbSets, overrides to the context. Without you'll probably need to register the generic overload by yourself to make it injectable to the constructor or use the interface and write an own database context inherited from IConfigurationDbContext