Add windows event log sink
This allows to step through multiple configured sinks in the DefaultEventService. For instance writing events to windows event log, or to databases. I've added the windows event log sink. As an idea it would be nice to write the logs to the grants db as another persistent storage besides the windows implementation.

To write data into event log, service must either run as administrator or an registry key must be added