Rename RemoveExpiredGrantsAsync
Changed to CleanupGrantsAsync. This is more accurate, because we are
performing all cleanup tasks within this method, which might include removing
consumed grants and expired device codes in addition to expired grants.

This is strictly speaking a breaking change, but it a safe one to make
because even though this method is public, it is not virtual. The
assumption is that the only users of this class are
- IdentityServer itself, which we change during the rename
- Classes that derive from TokenCleanupService, which can't override
this non-virtual method