Filter authorization codes from token request logs
This changes the default TokenRequestSensitiveValuesFilter to
include the code parameter. This is for consistency with the parsed
value, which is already being obfuscated. Previous logs that included
codes are not really a concern, since codes are one time use only - by
the time they are logged, they are not likely to be useful anymore.