Fix log messages in backchannel user validator
This fixes #1345, which happened because the passed in log message
contained a placeholder, and LogWithRequestDetails also added a
placeholder for the values, resulting in too many placeholders.

The fix is to remove the extra placeholder from LogWithRequestDetails and
review all calls to it to ensure the passed log message always contains
placeholders for the values.