Fix types when reading jwk from DPoP proof token
The new wilson version's limitation when reading a complex
Header and Payload value into a dictionary is that it requires
that you specify the concrete Dictionary type rather than the
IDictionary interface.

Thus, we can go back to the original way of reading the jwk,
and then serializing it into json to pass to the JsonWebKey
ctor, we just need to use Dictionary instead of IDictionary
when retrieving the header.