Remove a review note
OIDC requires that when request objects are used, the client pass the
client id and response type, in order to make the request a valid oauth
request. With PAR, we don't need to worry about that, since PAR is an
OAuth specification, and the requests we're making are valid PAR
requests. And in fact, an example in RFC 9126 pushes a request object
containing a response_type claim without duplicating the response type
outside the request object.

Since OIDC is asking for this duplication in other circumstances in
order to be a valid oauth request, and the relevant oauth specification
doesn't require it in this circumstance, I'm leaving it out.